+++ apps/frontend/src/App.jsx
@@ -2,16 +2,25 @@ import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './App.css'
-import encryptData  from 'encrypt-handler';
+import index  from 'encrypt-handler';
+const { encryptData, decryptData, hashData } = index;
+// import { encryptData, decryptData, hashData } from 'encrypt-handler';
+
   console.log(encryptData);
+    console.log(decryptData);
 function App() {
   const [count, setCount] = useState(0)
 
-//  const encryptedData= encryptData("samresh","samresh" );
+const encryptedData= encryptData("samresh","samresh" );
+const decryptedData = decryptData(
+  "U2FsdGVkX1+8j6OBkcJmRUFN4uWY+XHnVbWvMaAX1LU=",
+  "samresh"
+);
   return (
     <>
       <div>
-        {/* <h2>{encryptedData}</h2> */}
+        <h2>{encryptedData}</h2>
+        <h2>{decryptedData}</h2>
         <a href="https://vitejs.dev" target="_blank" rel="noreferrer">
           <img src={viteLogo} className="logo" alt="Vite logo" />
         </a>
