--- packages/encrypt-handler/index.cjs
@@ -1,3 +1,0 @@
-const handlers = require("./encryptDecryptHandler.js")
-
-module.exports = handlers;
