--- packages/encrypt-handler/index.mjs
@@ -1,4 +1,0 @@
-// import { encryptData, decryptData, hashData } from "./encryptDecryptHandler.js";
-import handlers from "./encryptDecryptHandler.js";
-// import handlers from "./encryptDecryptHandler.js";
-export default { ...handlers };
