+++ packages/encrypt-handler/src/index.js
@@ -1,0 +1,8 @@
+const  { encryptData,
+  decryptData,
+  hashData} = require("./encryptDecryptHandler")
+module.exports = {
+  encryptData,
+  decryptData,
+  hashData,
+};
