+++ packages/encrypt-handler/unbuild.config.mjs
@@ -1,0 +1,11 @@
+
+export default {
+  builds: {
+    esm: {
+      entryPoints: ["./src/index.js"],
+    },
+    cjs: {
+      entryPoints: ["./src/index.js"],
+    },
+  },
+};
